refactor LocalRepo object to own file and test
